requirements.txt
loguru
tqdm
pyyaml
pandas
matplotlib
scikit-learn
scikit-image
scipy
opencv-python
seaborn
albumentations
tabulate
warmup-scheduler
einops
pthflops
torchsummary
thop
segmentation-mask-overlay==0.3.4
